Moved DatabaseTest to the tests/fixtures folder
