+++ tests/api/mockapi/axis.py
@@ -2,8 +2,8 @@ from api.axis import Axis360API
 from core.model import get_one_or_create
 from core.model.collection import Collection
 from core.model.configuration import ExternalIntegration
-from core.testing import DatabaseTest
 from core.util.http import HTTP
+from tests.fixtures.db import DatabaseTest
 
 
 class MockAxis360API(Axis360API):
