+++ tests/api/mockapi/odl.py
@@ -1,8 +1,9 @@
 from api.odl import SharedODLAPI
 from core.model import get_one_or_create
 from core.model.collection import Collection
-from core.testing import DatabaseTest, MockRequestsResponse
+from core.testing import MockRequestsResponse
 from core.util.http import HTTP
+from tests.fixtures.db import DatabaseTest
 
 
 class MockSharedODLAPI(SharedODLAPI):
