+++ tests/test_s3_analytics_provider.py
@@ -16,7 +16,7 @@ from core.model import (
     create,
 )
 from core.s3 import S3Uploader, S3UploaderConfiguration
-from core.testing import DatabaseTest
+from tests.fixtures.db import DatabaseTest
 
 
 class TestS3AnalyticsProvider(DatabaseTest):
